+++ js/react/es5/DataFormsJS.js
@@ -492,7 +492,8 @@ function IsLoaded(props) {
   return React.cloneElement(props.children, {
     data: props.data,
     params: props.params,
-    handleChange: props.handleChange
+    handleChange: props.handleChange,
+    childProps: props.childProps
   });
 }
 
@@ -771,7 +772,8 @@ var JsonData = function (_React$Component3) {
         fetchState: this.state.fetchState,
         data: this.state.data,
         params: this.state.params,
-        handleChange: this.handleChange
+        handleChange: this.handleChange,
+        childProps: this.props.childProps
       }, this.props.isLoaded));
     }
   }]);